refactor: order-service 스키마 전면 개편 및 이벤트 계약 skuId 타입 변경
Order 도메인 모델을 DDL 스키마에 맞춰 전면 재구성:
- 테이블명 변경 (orders → order_order, order_items → order_item)
- 금액 모델 변경 (subtotal/discount/shippingFee/total)
- 배송지를 JSON 스냅샷으로 전환, pricingSnapshot 추가
- orderNo, currency, placedAt/paidAt/canceledAt 필드 추가
- OrderStatusHistory 상태 전이 이력 테이블 추가
- OrderIdempotency API 멱등성 테이블 추가
- INIT 상태 및 환불 로직 전면 제거
- skuId 타입 String → Long 변경 (이벤트 계약 포함)

+++ services/order-service/src/main/kotlin/com/koosco/orderservice/infra/messaging/kafka/consumer/KafkaStockConfirmedConsumer.kt
@@ -65,7 +65,7 @@ class KafkaStockConfirmedConsumer(
                     orderId = stockConfirmedEvent.orderId,
                     reservationId = stockConfirmedEvent.reservationId,
                     items = stockConfirmedEvent.items.map {
-                        MarkOrderConfirmedCommand.MarkedConfirmedItem(it.skuId, it.quantity)
+                        MarkOrderConfirmedCommand.MarkedConfirmedItem(it.skuId.toLong(), it.quantity)
                     },
                 ),
             )
